Fix CI compile round 1: GeminiOAuthManager.login companion + MessageDigest.update(Int)
- GeminiOAuthManager.login was an instance method (doc said static) — moved
  into companion object mirroring ClaudeOAuthManager/OpenAIOAuthManager so
  AddProviderScreen/ProviderDetailScreen static call sites resolve.
- TelegramClient tee stream: MessageDigest.update has no Int overload —
  use b.toByte() in single-byte write().

+++ src/android/app/src/main/java/com/openminis/app/backup/remote/TelegramClient.kt
@@ -446,7 +446,7 @@ class TelegramClient(@Suppress("unused") private val context: Context) {
             val tee = object : java.io.OutputStream() {
                 override fun write(b: Int) {
                     out.write(b)
-                    digest.update(b)
+                    digest.update(b.toByte())
                 }
                 override fun write(b: ByteArray, off: Int, len: Int) {
                     out.write(b, off, len)
